requirements.txt
litellm>=1.0.0
tenacity>=8.2.0
python-dotenv>=0.19.0
jsonschema>=4.0.0